requirements/loca.txt
-r base.txt

flake8==3.8.3
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
pytest-django==3.9.0